test(vsock-guest): split exec-control tests by behavior
