+++ crates/vsock-guest/src/exec_control/tests/mod.rs
@@ -1,0 +1,5 @@
+mod forwarding;
+mod handle;
+mod registry;
+mod sink_lifecycle;
+mod support;
